organized utility package and deleted files that will become part of a plugin
