+++ src/main/scala/de/tubs/cs/ias/plotalyzer/utility/output/Time.scala
@@ -1,4 +1,4 @@
-package de.tubs.cs.ias.plotalyzer.utility
+package de.tubs.cs.ias.plotalyzer.utility.output
 
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
