+++ version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "1.0.0"
+ThisBuild / version := "1.0.1"
